Start extracting fetch tasks to test
