+++ crates/kitsune_p2p/kitsune_p2p/src/spawn/actor/fetch/mod.rs
@@ -1,0 +1,4 @@
+mod fetch_task;
+mod response_config;
+
+pub use response_config::FetchResponseConfig;
