+++ crates/kitsune_p2p/kitsune_p2p/src/spawn/actor/space/bootstrap_task.rs
@@ -164,7 +164,7 @@ impl BootstrapTask {
 
 #[cfg(test)]
 mod tests {
-    use crate::event::{KitsuneP2pEvent, PutAgentInfoSignedEvt};
+    use crate::event::PutAgentInfoSignedEvt;
     use crate::fixt::AgentInfoSignedFixturator;
     use crate::fixt::KitsuneSpaceFixturator;
     use crate::spawn::actor::space::bootstrap_task::{BootstrapService, BootstrapTask};
@@ -174,6 +174,7 @@ mod tests {
         SpaceInternalHandler, SpaceInternalHandlerResult, VecMXM, WireConHnd,
     };
     use crate::spawn::actor::MetaNetCon;
+    use crate::spawn::test_util::HostStub;
     use crate::types::actor::BroadcastData;
     use crate::wire::Wire;
     use crate::KitsuneP2pResult;
@@ -701,69 +702,6 @@ mod tests {
             }
 
             async move { Ok(self.agents.clone()) }.boxed().into()
-        }
-    }
-
-    struct HostStub {
-        respond_with_error: Arc<AtomicBool>,
-        put_events: Receiver<PutAgentInfoSignedEvt>,
-        abort_handle: AbortHandle,
-    }
-
-    impl HostStub {
-        fn start(mut host_receiver: Receiver<KitsuneP2pEvent>) -> Self {
-            let (mut sender, receiver) = channel(10);
-
-            let respond_with_error = Arc::new(AtomicBool::new(false));
-            let handle = tokio::spawn({
-                let task_respond_with_error = respond_with_error.clone();
-                async move {
-                    while let Some(evt) = host_receiver.next().await {
-                        match evt {
-                            KitsuneP2pEvent::PutAgentInfoSigned { input, respond, .. } => {
-                                println!("Responding to requests");
-
-                                if task_respond_with_error.load(Ordering::SeqCst) {
-                                    respond.respond(Ok(async move {
-                                        Err(KitsuneP2pError::other("a test error"))
-                                    }
-                                    .boxed()
-                                    .into()));
-                                    continue;
-                                }
-
-                                sender.send(input).await.unwrap();
-                                respond.respond(Ok(async move { Ok(()) }.boxed().into()));
-                            }
-                            _ => panic!("Unexpected event - {:?}", evt),
-                        }
-                    }
-                }
-            });
-
-            HostStub {
-                respond_with_error,
-                put_events: receiver,
-                abort_handle: handle.abort_handle(),
-            }
-        }
-
-        async fn next_event(&mut self, timeout: Duration) -> PutAgentInfoSignedEvt {
-            tokio::time::timeout(timeout, self.put_events.next())
-                .await
-                .unwrap()
-                .unwrap()
-        }
-
-        async fn try_next_event(
-            &mut self,
-            timeout: Duration,
-        ) -> Result<Option<PutAgentInfoSignedEvt>, Elapsed> {
-            tokio::time::timeout(timeout, self.put_events.next()).await
-        }
-
-        fn abort(&self) {
-            self.abort_handle.abort();
         }
     }
 }
