+++ crates/kitsune_p2p/kitsune_p2p/src/spawn/actor/test_util/mod.rs
@@ -1,5 +1,7 @@
+mod host_stub;
 mod internal_stub;
 mod space_internal_stub;
 
+pub use host_stub::HostStub;
 pub use internal_stub::InternalStub;
 pub use space_internal_stub::SpaceInternalStub;
